requirements.txt
Django
django-jazzmin
python-decouple
Pillow
mysqlclient
mysql-connector-python
